Add crt maintenance view
