--- packages/atlas/src/views/studio/CrtView/index.ts
@@ -1,1 +1,0 @@
-export * from './CrtView'
